src/main/java/service/impl/AuthenticationServiceImpl.java
package service.impl;

import exception.AuthenticationException;
import model.User;
import service.AuthenticationService;
import service.UserService;
import java.util.Optional;

public class AuthenticationServiceImpl implements AuthenticationService {
    private final UserService userService;

    public AuthenticationServiceImpl(UserService userService) {
        this.userService = userService;
    }

    @Override
    public User register(String login, String password) {
        User user = new User();
        user.setLogin(login);
        user.setPassword(password);
        return userService.save(user);
    }

    @Override
    public User login(String login, String password) throws AuthenticationException {
        Optional<User> user = userService.findByLogin(login);
        if (user.isEmpty() || !user.get().getPassword().equals(password)) {
            throw new AuthenticationException("Login or password is invalid");
        }
        return user.get();
    }
}
